fix 파일 및 함수 수정
