+++ src/apis/postApi.js
@@ -3,7 +3,6 @@ import axios from 'axios';
 import { serverURL, POST } from './endpoints';
 import { getCookie } from '../utils/cookie';
 
-
 export const getPostByChannelApi = async channelId => {
   try {
     const response = await devAPI.get(POST.getChannelPost(channelId));
@@ -78,3 +77,13 @@ export const createPostApi = async formData => {
   }
 };
 
+export const getUserPost = async userId => {
+  try {
+    const response = await devAPI.get(POST.getAuthorPost(userId));
+
+    return response.data;
+  } catch (error) {
+    console.error('유저의 포스트를 불러오지 못했습니다.', error);
+    throw new Error('채널 데이터를 불러오는 데 실패했습니다.');
+  }
+};
